+++ Seguradora/Models/Veiculo.cs
@@ -39,11 +39,13 @@ namespace Seguradora.Models
             idVeiculo = pIdVeiculo;
             placa = pPlaca;
             ano = pAno;
+            aModelo = pModelo;
         }
 
         public Modelo AModelo
         {
             get { return aModelo; }
+            set { aModelo = value; }
         }
 
         public int Ano
@@ -72,6 +74,9 @@ namespace Seguradora.Models
             if (retorno.Length != 0)
                 return retorno;
 
+            if (ValidarIdExistente())
+                return "Id de veiculo já utilizada\r\n";
+
             string cmdInserir = $"INSERT INTO veiculo(IdVeiculo, IdModelo, Placa, Ano) VALUES({IdVeiculo}, {aModelo.IdModelo}, '{placa}', {ano});";
             return ConexaoBD.GetInstance().ExecutarSql(cmdInserir);
         }
@@ -81,6 +86,9 @@ namespace Seguradora.Models
             string retorno = Validar();
             if (retorno.Length != 0)
                 return retorno;
+
+            if (!ValidarIdExistente())
+                return "Veiculo com este ID não existe.\r\n";
 
             string cmdAlterar = $"UPDATE veiculo SET IdModelo={aModelo.IdModelo}, Placa='{placa}', Ano={ano} WHERE IdVeiculo={IdVeiculo};";
             return ConexaoBD.GetInstance().ExecutarSql(cmdAlterar);
@@ -92,6 +100,9 @@ namespace Seguradora.Models
             if (retorno.Length != 0)
                 return retorno;
 
+            if (!ValidarIdExistente())
+                return "Veiculo com este ID não existe.\r\n";
+
             string cmdExcluir = $"DELETE FROM veiculo WHERE IdVeiculo={idVeiculo};";
             return ConexaoBD.GetInstance().ExecutarSql(cmdExcluir);
         }
@@ -101,6 +112,9 @@ namespace Seguradora.Models
             string retorno = Utils.ValidarId(idVeiculo, "veiculo");
             if (retorno.Length != 0)
                 return retorno;
+
+            if (!ValidarIdExistente())
+                return "Veiculo com este ID não existe.\r\n";
 
             string cmdRecuperar = $"SELECT * FROM veiculo WHERE IdVeiculo={idVeiculo};";
             DataTable tabela = ConexaoBD.GetInstance().ExecutarQuery(cmdRecuperar, out retorno);
@@ -116,10 +130,6 @@ namespace Seguradora.Models
                 ano = Convert.ToInt32(linha["Ano"]);
                 aModelo = new Modelo(Convert.ToInt32(linha["IdModelo"]));
                 retorno = "";
-            }
-            else
-            {
-                retorno = "Nenhum veiculo encontrado. ";
             }
 
             return retorno;
@@ -150,6 +160,32 @@ namespace Seguradora.Models
         }
         #endregion
 
+        private bool ValidarIdExistente()
+        {
+            string retorno;
+            string cmdSelectId = $"SELECT idVeiculo FROM veiculo WHERE idVeiculo={IdVeiculo}";
+            DataTable tabela = ConexaoBD.GetInstance().ExecutarQuery(cmdSelectId, out retorno);
+
+            return tabela.Rows.Count != 0;
+        }
+
+        private string ValidarModelo()
+        {
+            string retorno = "";
+            string auxRetorno = "";
+            string cmdSelectModelo = $"SELECT idModelo FROM modelo WHERE idModelo={aModelo.IdModelo};";
+
+            DataTable tabela = ConexaoBD.GetInstance().ExecutarQuery(cmdSelectModelo, out retorno);
+
+            if (retorno.Length != 0)
+                return retorno;
+
+            if (tabela.Rows.Count == 0)
+                auxRetorno += "Id de modelo não cadastrado!\r\n";
+
+            return auxRetorno;
+        }
+
         private string Validar()
         {
             {
@@ -157,16 +193,14 @@ namespace Seguradora.Models
                 retorno += Utils.ValidarId(idVeiculo, "veiculo");
 
                 if (placa.Length == 0)
-                {
-                    retorno += "placa não pode ser vazia.\r\n";
-                }
+                    retorno += "Placa não pode ser vazia.\r\n";
+                if (placa.Length > 45)
+                    retorno += "Placa maior que o limite de 45 caracteres.\r\n";
 
                 if (ano < 0)
-                {
                     retorno += "ano não pode ser negativo.\r\n";
-                }
 
-                return retorno;
+                return retorno += ValidarModelo();
             }
         }
     }
